Upgrade to work with iced mainline.
There have been some significant breaking changes in iced, in particular a large refactoring of the crates themselves (many namespaces have changed) as well as some small but important changes to the wgpu renderer, which now uses the render queue instead of `StagingBelt`.

This currently depends on an `iced` fork that downgrades back to wgpu 0.15.1, since Bevy has not yet put out an official release for the 0.16 update. So we can either mainline _everything_ (and risk instability) or temporarily used this `iced` branch until either Bevy and Iced are back in sync or until it is no longer possible to keep the branch in sync.