Clean up imports.
Most of the files are already well-organized and using single-name imports for all important types, but `lib.rs` wasn't. Aside from improving readability, this refactoring should make it easier to keep up with changes happening in `iced` itself, i.e. so that only the import needs to be updated and not dozens of lines in the implementation.